vgshop_backend/requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
django-embed-video==1.4.10
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.13
pillow==12.2.0
PyJWT==2.12.1
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
